Changed all prefix commands to SlashCommands  - Added BotRoleType and CommandRequestType  - Cleaned up the code  - Fixed small bug at RKICorona

+++ src/main/java/de/moinFT/utils/DBServerArray.java
@@ -108,7 +108,7 @@ public class DBServerArray {
         return this.ServerID;
     }
 
-    public long getCommandTimeout() {
+    public int getCommandTimeout() {
         return this.CommandTimeout;
     }
